fix a small error in rpn result

+++ cpp09/ex01/RPN.hpp
@@ -38,7 +38,7 @@ class	RPN {
 	};
 
 		private:
-	int				_result;
+	long long		_result;
 	int				_operand1;
 	int				_operand2;
 	char			_operator;
